backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
anthropic>=0.54.0
openai>=1.0.0
pinecone==7.2.0
pandas==2.2.3
python-dotenv==1.0.0
gunicorn==21.2.0
pdfplumber>=0.11.0
fpdf2>=2.8.0
pytest==7.4.4
pytest-flask==1.3.0